feat: add support for `dubcap-time` option in mission command
The `dubcap-time` option allows users to specify the time remaining for
the double capacity event. This information is used to calculate the launch
time of missions in relation to the end of the event. The `hhmmss` library
is added to parse the time format provided by users for the double capacity
event.