feat: Update Egg type to int32 and add EggName field
Changed the Egg field type in EggIncContract struct from int to int32 to
accommodate the updated data. Added a new EggName field to hold the name of
the Egg based on the egg index. Calculated and stored qTargetAmount by
dividing TargetAmount by 1e15. Updated coopAllowed logic to default to false
and check for nil value to prevent errors.