feat(boost): Add support for future token log times
This change adds support for tracking the time values associated with the
future token log values. Previously, only the log values were tracked,
but now the corresponding time values are also stored and used when
calculating the token value.

This allows for more accurate representation of the token value over time,
as the time information is now available in addition to the log values.