feat(track): update duration display logic
The changes update the logic for displaying the duration of a token
tracking event. Previously, the duration was calculated based on the
difference between the start time and the time from the coop status.
This has been changed to use the pre-calculated duration time, which
provides a more accurate and consistent display of the duration.