feat(boost): change contract state check logic
The changes in this commit update the contract state check logic in the `boost_button_reactions.go` file. The previous check for `ContractStateBanker` and `ContractStateFastrun` has been replaced with a check for `ContractStateSignup`. This change ensures that the correct set of requestors is generated based on the contract state.

Additionally, in the `app_emoji.go` file, a TODO comment has been added to address an issue related to the unsupported protocol scheme when creating emojis in Discord.