feat(banners): Add support for multiple banner styles
This change adds support for generating multiple banner styles for a given
egg. The new styles include "chill", "aco", "fastrun", and "leaderboard".
Each style is represented by a unique image that is drawn in the lower left
corner of the banner. The code checks if any of the style images already
exist before generating them, and only generates the ones that don't exist.
This helps to improve performance and reduce unnecessary file I/O.