feat: Improve help message with contract playstyle links
The changes made in this commit improve the help message displayed to users when they interact with the "help" button in the Boost Bot. The main changes are:

1. The help message now includes a link to the relevant contract playstyle information, if the contract has a playstyle set. This allows users to easily access more details about the different playstyles supported by the bot.

2. The formatting of the help message has been improved by using a `strings.Builder` to construct the output string, which makes the code more readable and maintainable.

3. The cooldown time for updating the contract duration has been reduced from 3 minutes to 10 seconds, as this is a common operation that users may need to perform frequently.

chore: Reduce duration update cooldown to 10 seconds

The changes in this commit reduce the cooldown time for updating the contract duration from 3 minutes to 10 seconds. This is a common operation that users may need to perform frequently, so reducing the cooldown time will improve the user experience.