feat(timer): Add sticky timer functionality
This commit adds a new feature to the timer command, allowing users to set a
"sticky" timer that is remembered across command invocations. The changes
include:

- Adding a new `farmerstate` package to manage user-specific settings
- Updating the `/timer` command to check for a stored "timer" setting and use
  that as the default duration if present
- Storing the user-provided duration in the "timer" setting when a new timer is
  set
- Making the "duration" option optional, as the sticky timer can be used as the
  default

The goal of these changes is to provide a more convenient user experience,
allowing users to easily set a recurring timer without having to re-enter the
duration each time.