feat(boost): Add toggle-contract-pings command
This commit adds a new command `/toggle-contract-pings` that allows users to
toggle whether they receive pings for a contract they are part of. When
enabled, the user will be added to the contract's role, and when disabled,
the user will be removed from the role.

Additionally, this commit includes a change to the `DeleteContract` function
to also delete the contract's role when the contract is deleted.