feat(teamwork): Update directory for reading and writing files
The changes in this commit update the directory paths used for reading and writing files in the `teamwork.go` and `coop_status.go` files.

The main changes are:

1. In `teamwork.go`, the directory for reading files is changed from `"ttbb-data"` to `"ttbb-data/pb"`.
2. In `coop_status.go`, the directory for writing files is changed from `"ttbb-data"` to `"ttbb-data/pb"`.
3. Additionally, in `coop_status.go`, a new directory `"ttbb-data/pb"` is created if it doesn't already exist before writing the file.

These changes are made to organize the files in a more structured way, separating the protocol buffer (PB) files from other files that may be present in the `"ttbb-data"` directory.