feat: use nickname instead of full name in boost message
The changes in this commit update the boost message to use the nickname of the
booster instead of their full name. This makes the message more concise and
easier to read, especially for longer names. The `Nick` field is used instead
of the `Name` field to retrieve the booster's nickname.