feat(contract): set sink boost position based on contract duration
The changes made in this commit are focused on setting the sink boost position for a contract based on its estimated duration. For long contracts (10 hours or more), the sink boost position is set to `SinkBoostLast`, which means the sink will boost last. For short contracts (less than 10 hours), the sink boost position is set to `SinkBoostFirst`, which means the sink will boost first. This change is intended to improve the overall performance and efficiency of the contract system.