feat(deps): update google.golang.org/api and google.golang.org/protobuf
This commit updates the versions of the google.golang.org/api and google.golang.org/protobuf dependencies in the go.mod file. The changes are:

- google.golang.org/api updated from v0.247.0 to v0.248.0
- google.golang.org/protobuf updated from v1.36.7 to v1.36.8

These updates are necessary to ensure the project is using the latest versions of these dependencies, which may include bug fixes, security updates, or new features.