feat(boost): add progenitor support for contract creation
This commit adds support for specifying progenitor farmers when creating a new contract. The `AddFarmerToContract` function has been updated to accept a `progenitor` boolean parameter, which determines whether the added farmer should be treated as a progenitor or not.

The contract creation command has also been updated to include a new `progenitors` option, which allows the user to specify a list of mentions to seed farmers for the contract. The function will validate the mentioned users and only add valid progenitors to the contract.

This change allows contract organizers to easily set up a group of progenitor farmers to start boosting the contract, which can help kickstart the contract and encourage participation from other farmers.