feat(banners): Add automatic download of latest egg images
This change adds a check to see if the overlay image file exists before attempting to load it. If the file does not exist, it downloads the latest egg images from the GitHub repository.

The changes also include:
- Resizing the overlay image to a fixed size of 128x128 pixels using the `draw.NearestNeighbor.Scale` function.
- Adjusting the font size of the text overlay if the text is too wide to fit within the available space.
- Refactoring the `addLabel` function to use the adjusted font face.
- Adding a new `DownloadLatestEggImages` function to download the latest egg images from the GitHub repository.

These changes ensure that the latest egg images are always available and that the text overlay is properly sized to fit within the banner image.