feat: Update Google API dependencies
The changes in this commit update the Google API dependencies to the latest versions. Specifically, the `google.golang.org/genproto/googleapis/api` and `google.golang.org/genproto/googleapis/rpc` dependencies have been updated to version `20250826171959-ef028d996bc1`. This update is necessary to ensure compatibility with the latest Google API changes and to keep the project up-to-date.