feat(boost): remove CRT-related functionality
The changes in this commit remove the CRT-related functionality from the boost package. This includes the removal of the `reactionCRT` function, the `buttonReactionCheck`, `buttonReactionTruck`, `buttonReactionLeg`, `buttonReactionTango`, and `remove` functions, as well as the entire `state_crt.go` file.

These changes were made to simplify the codebase and remove functionality that is no longer needed or used in the current version of the application.